Added more propType structure
